Finish header section
The file CSS leverages media queries to make the header's layout
responsive. On the other hand, in the index.js new code was added
to give functionality to the button that must show the user
the options to go to the menu, contact and home once the screen
has a specific width.